backend/requirements.txt
# Web framework
Flask==3.0.0
Flask-CORS==4.0.0

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==5.1.0

# Reddit API (optional - can use scraping instead)
praw==7.7.1

# Google Gemini AI
google-generativeai==0.3.2

# Environment variables
python-dotenv==1.0.0

# Rate limiting
ratelimit==2.2.1

# Development
pytest==7.4.3
pytest-flask==1.3.0
